requirements.txt
babel==2.17.0
beautifulsoup4==4.12.3
breadability==0.1.20
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
courlan==1.3.2
cssselect==1.3.0
dateparser==1.2.2
docopt==0.6.2
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.19.1
greenlet==3.0.3
htmldate==1.9.3
idna==3.10
jieba3k==0.35.1
joblib==1.5.1
jusText==3.0.2
lxml==5.4.0
lxml_html_clean==0.4.2
markdown-it-py==4.0.0
mdurl==0.1.2
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.3.2
orjson==3.11.2
pillow==11.3.0
playwright==1.47.0
pycountry==24.6.1
pyee==12.0.0
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
readability-lxml==0.8.1
regex==2025.7.34
requests==2.32.3
requests-file==2.1.0
rich==14.1.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.7
sumy==0.11.0
tinysegmenter==0.3
tld==0.13.1
tldextract==5.3.0
tqdm==4.66.4
trafilatura==2.0.0
typer==0.12.5
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
